+++ models/products.js
@@ -29,7 +29,7 @@ exports.createProduct = async (data) => {
     created_at: createdAt,
     updated_at: createdAt
   });
-  console.log("query -->", query.toQuery())
+  console.info("query -->", query.toQuery())
   return query;
 };
 
@@ -47,7 +47,7 @@ exports.updateProduct = async (data) => {
     .where('id', data.id)
     .update(toBeUpdated);
 
-  console.log("query -->", query.toQuery())
+  console.info("query -->", query.toQuery())
   return query;
 };
 
@@ -58,7 +58,7 @@ exports.removeProduct = async (id) => {
       available: 0,
       updated_at: moment().format('YYYY-MM-DD HH:mm:ss')
     });
-  console.log("query -->", query.toQuery())
+  console.info("query -->", query.toQuery())
   return query;
 };
 
@@ -69,56 +69,9 @@ exports.reAddProduct = async (id) => {
       available: 1,
       updated_at: moment().format('YYYY-MM-DD HH:mm:ss')
     });
-  console.log("query -->", query.toQuery())
+  console.info("query -->", query.toQuery())
   return query;
 };
-
-exports.createSeller = async (data) => {
-  const createdAt = moment().format('YYYY-MM-DD HH:mm:ss');
-  const query = db.write('seller').insert({
-    user_id: data.user_id || null,
-    name: data.name || null,
-    about_us: data.about_us || null,
-    email: data.email || null,
-    phone: data.phone || null,
-    verified_phone: data.verified_phone || 0,
-    verified_email: data.verified_email || 0,
-    verified_account: data.verified_account || 0,
-    DOB: data.DOB || null,
-    street: data.street || null,
-    city: data.city || null,
-    zipcode: data.zipcode || null,
-    state: data.state || null,
-    country: data.country || null,
-    logo: data.logo || null,
-    created_at: createdAt,
-    updated_at: createdAt
-  });
-  console.log("query -->", query.toQuery())
-  return query;
-};
-
-exports.createWarehouse = async (data) => {
-  const createdAt = moment().format('YYYY-MM-DD HH:mm:ss');
-  const query = db.write('warehouse').insert({
-    seller_id: data.seller_id || null,
-    name: data.name || null,
-    about_us: data.about_us || null,
-    email: data.email || null,
-    phone: data.phone || null,
-    street: data.street || null,
-    city: data.city || null,
-    zipcode: data.zipcode || null,
-    state: data.state || null,
-    country: data.country || null,
-    pictures: data.pictures || null,
-    created_at: createdAt,
-    updated_at: createdAt
-  });
-  console.log("query -->", query.toQuery())
-  return query;
-};
-
 
 exports.createProductCategory = async (data) => {
   const createdAt = moment().format('YYYY-MM-DD HH:mm:ss');
@@ -127,6 +80,6 @@ exports.createProductCategory = async (data) => {
     created_at: createdAt,
     updated_at: createdAt
   });
-  console.log("query -->", query.toQuery())
+  console.info("query -->", query.toQuery())
   return query;
 };
