+++ routes/products.js
@@ -37,16 +37,6 @@ router.post('/readd/:id', authenticator, allowSeller, async (req, res) => {
   req.body.id = Number(req.params.id);
   const response = await prodController.reAddProduct(req.body)
   return res.status(response.status).send(response)
-})
-
-router.post('/seller', authenticator, async (req, res) => {
-  const response = await prodController.createSeller(req.body)
-  return res.status(response.status).send(response)
-})
-
-router.post('/warehouse', authenticator, allowAdmin, async (req, res) => {
-  const response = await prodController.createWarehouse(req.body)
-  return res.status(response.status).send(response)
-})
+});
 
 module.exports = router;
