+++ routes/users.js
@@ -62,9 +62,9 @@ router.post('/role', authenticator, allowAdmin, async (req, res) => {
   return res.status(response.status).send(response)
 });
 
-// router.post('/token',  async (req, res) => {
-//   const response = await userController.createUserToken(req.body);
-//   return res.status(response.status).send(response)
-// });
+router.post('/seller', authenticator, async (req, res) => {
+  const response = await userController.createSeller(req.body)
+  return res.status(response.status).send(response)
+})
 
 module.exports = router;
