+++ validators/products.js
@@ -28,31 +28,3 @@ exports.validateProductCategory = body => {
 
   return bodyStruct;
 };
-
-exports.validateSeller = body => {
-  const bodyStruct = {};
-  const arr = ['user_id', 'about_us', 'logo', 'name', 'verified', 'email', 'phone', 'street',  'city', 'state',
-  'country' ];
-
-  arr.map((item) => {
-    const check = body.hasOwnProperty(item);
-    if (!check) throw errorResponse(400, item+'Missing');
-    bodyStruct[item] = body[item];
-  });
-
-  return bodyStruct;
-};
-
-exports.validateWarehouse = body => {
-  const bodyStruct = {};
-  const arr = ['seller_id', 'pictures', 'about_us', 'verified', 'email', 'phone', 'street',  'city', 'state',
-  'country' ]
-
-  arr.map((item) => {
-    const check = body.hasOwnProperty(item);
-    if (!check) throw errorResponse(400, item+'Missing');
-    bodyStruct[item] = body[item];
-  });
-
-  return bodyStruct;
-};
