+++ validators/users.js
@@ -86,3 +86,17 @@ exports.validateAuth = body => {
   return bodyStruct;
 };
 
+exports.validateSeller = body => {
+  const bodyStruct = {};
+  const arr = ['user_id', 'about_us', 'logo', 'name', 'verified', 'email', 'phone', 'street',  'city', 'state',
+  'country' ];
+
+  arr.map((item) => {
+    const check = body.hasOwnProperty(item);
+    if (!check) throw errorResponse(400, item+'Missing');
+    bodyStruct[item] = body[item];
+  });
+
+  return bodyStruct;
+};
+
